+++ src/sanmar_sdk/__init__.py
@@ -1,0 +1,5 @@
+"""A wrapper for SanMar's API."""
+
+from .client import SanMar
+
+__all__ = ["SanMar"]