+++ docs/source/changelog.rst
@@ -1,2 +1,5 @@
 Changelog
 =========
+
+- :release:`0.1.0 <7th October 2023>`
+- :feature:`1` Initialize package